+++ Bookstore/src/main/Inventory.java
@@ -3,8 +3,8 @@ package main;
 import java.io.Serializable;
 
 public interface Inventory extends Serializable{
-	public double addBook(Book toAdd);
-	public double sellBook(Book toSell);
+	public int addBook(Book toAdd);
+	public int sellBook(Book toSell);
 	public double updatePrice(Book toUpdate);
 	public int getQuantityByTitle(String toFind);
 	public int getQuantityByID(Integer toFind);
